qemu_v8: Changes for using vhost-user-rpmb backend
1. Option added while invoking QEMU
2. Additional config options for linux build
3. Disable RPMB_EMU for optee-client
3. Change permissions for /dev/rpmb ro allow tee-supplicant to open it
However there is still issue with this as driver expects the process to
have CAP_SYS_RAWIO capability which is present onlty if we run
tee-supplicant as a root process.
Tried to change the capability of teh tee-supplicant by using set-cap,
so enabled the same in buildroot. However, this doesn;t work. So for now
to test it, from linux prompt :
killall tee-supplicant
tee-supplicant &

Then try xtest.

Signed-off-by: Ruchika Gupta <ruchika.gupta@linaro.org>